refactoring: move implementation details out off WalletManager

+++ src/x1d/Obsidian.Features.X1Wallet/Models/Wallet/MemoryPoolMetadata.cs
@@ -21,7 +21,7 @@ namespace Obsidian.Features.X1Wallet.Models.Wallet
 
         public string ConsensusError;
         public string MemoryPoolError;
-        public uint TransactionTime;
+        public long TransactionTime;
 
         #region overrides of Equals, GetHashCode, ==, !=
